move: Custom MDEditor => src/components/common
