--- src/app/(home)/(upsert)/_components/edit/editform/FormContentArea.tsx
@@ -1,19 +1,0 @@
-import { Dispatch } from 'react';
-
-import CustomMDEditor from '../../CustomMDEditor';
-
-type FormContentAreaProps = {
-  content: string;
-  setContent: Dispatch<React.SetStateAction<string>>;
-};
-
-const FormContentArea = ({ content, setContent }: FormContentAreaProps) => {
-  return (
-    <div className="">
-      <h5 className="block mb-2 text-gray-900 text-h5 font-bold">본문</h5>
-      <CustomMDEditor content={content} setContent={setContent} />
-    </div>
-  );
-};
-
-export default FormContentArea;
